Give autocompleters custom map outlets to avoid confusion
