+++ app/javascript/controllers/map_controller.js
@@ -558,6 +558,9 @@ export default class extends GeocodeController {
       // this.showBoxBtnTarget.disabled = false
     }
     this.dispatch("reenableBtns")
+    // FIXME: This is a problem when there are two maps on the page.
+    // It emits the event to both maps, but only one should respond.
+    // Button should have an action instead that only affects the map it's in.
     this.dispatchPointChanged({ lat: null, lng: null })
   }
