+++ app/views/controllers/observations/form/_details.html.erb
@@ -74,6 +74,7 @@ t_s = {
           create_text: :form_observations_create_locality.l,
           keep_text: :form_observations_use_locality.l,
           edit_text: :form_observations_edit_locality.l,
+          map_outlet: "#observation_form_map",
           # find_text: :form_locations_find_on_map.l,
           data: {
             map_target: "placeInput",
